+++ src/main/java/com/sfjs/crud/svc/BusinessService.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Service;
 import com.sfjs.conv.AccountConverter;
 import com.sfjs.conv.BusinessConverter;
 import com.sfjs.jpa.entity.BusinessEntity;
-import com.sfjs.crud.repo.BaseRepository;
+import com.sfjs.jpa.repo.BaseRepository;
 import com.sfjs.crud.response.BusinessResponse;
 
 import jakarta.transaction.Transactional;
