+++ src/main/java/com/sfjs/crud/svc/FellowService.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Service;
 import com.sfjs.conv.AccountConverter;
 import com.sfjs.conv.FellowConverter;
 import com.sfjs.jpa.entity.FellowEntity;
-import com.sfjs.crud.repo.BaseRepository;
+import com.sfjs.jpa.repo.BaseRepository;
 import com.sfjs.crud.response.FellowResponse;
 
 import jakarta.transaction.Transactional;
