+++ src/main/java/com/sfjs/gql/svc/CheckoutService.java
@@ -20,10 +20,10 @@ import com.sfjs.jpa.entity.FellowEntity;
 import com.sfjs.jpa.entity.NumericMetricEntity;
 import com.sfjs.jpa.entity.PaymentEntity;
 import com.sfjs.jpa.entity.PaymentStatus;
-import com.sfjs.crud.repo.BusinessRepository;
-import com.sfjs.crud.repo.FellowRepository;
-import com.sfjs.crud.repo.NumericMetricRepository;
-import com.sfjs.crud.repo.PaymentRepository;
+import com.sfjs.jpa.repo.BusinessRepository;
+import com.sfjs.jpa.repo.FellowRepository;
+import com.sfjs.jpa.repo.NumericMetricRepository;
+import com.sfjs.jpa.repo.PaymentRepository;
 import com.sfjs.crud.response.PaymentResponse;
 import com.sfjs.crud.svc.BusinessService;
 import com.sfjs.crud.svc.FellowService;
