+++ src/main/java/com/sfjs/gql/svc/JobApplicationService.java
@@ -16,9 +16,9 @@ import com.sfjs.jpa.entity.FellowEntity;
 import com.sfjs.jpa.entity.JobApplicationEntity;
 import com.sfjs.jpa.entity.JobApplicationNoteEntity;
 import com.sfjs.jpa.entity.JobListingEntity;
-import com.sfjs.crud.repo.JobApplicationNoteRepository;
-import com.sfjs.crud.repo.JobApplicationRepository;
-import com.sfjs.crud.repo.JobListingRepository;
+import com.sfjs.jpa.repo.JobApplicationNoteRepository;
+import com.sfjs.jpa.repo.JobApplicationRepository;
+import com.sfjs.jpa.repo.JobListingRepository;
 import com.sfjs.data.JobApplicationData;
 import com.sfjs.data.JobApplicationNoteData;
 import com.sfjs.security.AuthorizationService;
