+++ src/main/java/com/sfjs/gql/svc/JobListingService.java
@@ -17,9 +17,9 @@ import com.sfjs.jpa.entity.BaseEntity;
 import com.sfjs.jpa.entity.BusinessEntity;
 import com.sfjs.jpa.entity.InterviewProcessEntity;
 import com.sfjs.jpa.entity.JobListingEntity;
-import com.sfjs.crud.repo.BaseRepository;
-import com.sfjs.crud.repo.InterviewProcessRepository;
-import com.sfjs.crud.repo.JobListingRepository;
+import com.sfjs.jpa.repo.BaseRepository;
+import com.sfjs.jpa.repo.InterviewProcessRepository;
+import com.sfjs.jpa.repo.JobListingRepository;
 import com.sfjs.data.HybridDetailsData;
 import com.sfjs.data.InterviewProcessData;
 import com.sfjs.data.JobListingData;
