+++ src/main/java/com/sfjs/gql/svc/SignupService.java
@@ -28,21 +28,21 @@ import com.sfjs.jpa.entity.FellowEntity;
 import com.sfjs.jpa.entity.HobbyEntity;
 import com.sfjs.jpa.entity.LinkEntity;
 import com.sfjs.jpa.entity.RoleEntity;
-import com.sfjs.crud.repo.AccomplishmentRepository;
-import com.sfjs.crud.repo.AccountRepository;
-import com.sfjs.crud.repo.AwardRepository;
-import com.sfjs.crud.repo.BaseRepository;
-import com.sfjs.crud.repo.BookOrQuoteRepository;
-import com.sfjs.crud.repo.BusinessProfileRepository;
-import com.sfjs.crud.repo.BusinessRepository;
-import com.sfjs.crud.repo.EducationRepository;
-import com.sfjs.crud.repo.ExperienceLevelRepository;
-import com.sfjs.crud.repo.ExperienceRepository;
-import com.sfjs.crud.repo.FellowRepository;
-import com.sfjs.crud.repo.HobbyRepository;
-import com.sfjs.crud.repo.LinkRepository;
-import com.sfjs.crud.repo.ProfileRepository;
-import com.sfjs.crud.repo.RoleRepository;
+import com.sfjs.jpa.repo.AccomplishmentRepository;
+import com.sfjs.jpa.repo.AccountRepository;
+import com.sfjs.jpa.repo.AwardRepository;
+import com.sfjs.jpa.repo.BaseRepository;
+import com.sfjs.jpa.repo.BookOrQuoteRepository;
+import com.sfjs.jpa.repo.BusinessProfileRepository;
+import com.sfjs.jpa.repo.BusinessRepository;
+import com.sfjs.jpa.repo.EducationRepository;
+import com.sfjs.jpa.repo.ExperienceLevelRepository;
+import com.sfjs.jpa.repo.ExperienceRepository;
+import com.sfjs.jpa.repo.FellowRepository;
+import com.sfjs.jpa.repo.HobbyRepository;
+import com.sfjs.jpa.repo.LinkRepository;
+import com.sfjs.jpa.repo.ProfileRepository;
+import com.sfjs.jpa.repo.RoleRepository;
 import com.sfjs.data.AccomplishmentData;
 import com.sfjs.data.AwardData;
 import com.sfjs.data.BaseProfileData;
