+++ src/main/java/com/sfjs/jpa/repo/AddressRepository.java
@@ -1,11 +1,11 @@
-package com.sfjs.crud.repo;
+package com.sfjs.jpa.repo;
 
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.sfjs.jpa.entity.RoleEntity;
+import com.sfjs.jpa.entity.AddressEntity;
 
 @Repository
 @Transactional
-public interface RoleRepository extends BaseRepository<RoleEntity> {
+public interface AddressRepository extends BaseRepository<AddressEntity> {
 }
