+++ src/main/java/com/sfjs/jpa/repo/BookOrQuoteRepository.java
@@ -1,4 +1,4 @@
-package com.sfjs.crud.repo;
+package com.sfjs.jpa.repo;
 
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
