+++ src/main/java/com/sfjs/jpa/repo/HobbyRepository.java
@@ -1,11 +1,12 @@
-package com.sfjs.crud.repo;
+package com.sfjs.jpa.repo;
 
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.sfjs.jpa.entity.AddressEntity;
+import com.sfjs.jpa.entity.HobbyEntity;
 
 @Repository
 @Transactional
-public interface AddressRepository extends BaseRepository<AddressEntity> {
+public interface HobbyRepository extends BaseRepository<HobbyEntity> {
+
 }
