+++ src/main/java/com/sfjs/jpa/repo/PaymentRepository.java
@@ -1,12 +1,11 @@
-package com.sfjs.crud.repo;
+package com.sfjs.jpa.repo;
 
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.sfjs.jpa.entity.LinkEntity;
+import com.sfjs.jpa.entity.PaymentEntity;
 
 @Repository
 @Transactional
-public interface LinkRepository extends BaseRepository<LinkEntity> {
-
+public interface PaymentRepository extends BaseRepository<PaymentEntity> {
 }
