+++ src/main/java/com/sfjs/jpa/repo/RoleRepository.java
@@ -1,12 +1,11 @@
-package com.sfjs.crud.repo;
+package com.sfjs.jpa.repo;
 
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.sfjs.jpa.entity.HobbyEntity;
+import com.sfjs.jpa.entity.RoleEntity;
 
 @Repository
 @Transactional
-public interface HobbyRepository extends BaseRepository<HobbyEntity> {
-
+public interface RoleRepository extends BaseRepository<RoleEntity> {
 }
