+++ src/main/java/com/sfjs/security/AuthorizationService.java
@@ -16,8 +16,8 @@ import org.springframework.stereotype.Service;
 
 import com.sfjs.jpa.entity.AccountEntity;
 import com.sfjs.jpa.entity.ResetPasswordTokenEntity;
-import com.sfjs.crud.repo.AccountRepository;
-import com.sfjs.crud.repo.ResetPasswordTokenRepository;
+import com.sfjs.jpa.repo.AccountRepository;
+import com.sfjs.jpa.repo.ResetPasswordTokenRepository;
 import com.sfjs.gql.schema.Result;
 
 import jakarta.servlet.http.HttpServletResponse;
